+++ CHANGELOG.md
@@ -1,10 +1,15 @@
+2.0.1
+=====
+
+* (improvement) Improve exception message for better tracing/debugging when hydration of a `Story` fails due to invalid data.
+
+
 2.0.0
 =====
 
 * (improvement) Add `StoryInterface`. 
 * (bc) Use `fullSlug` in `StoryLinkData`. 
-* (bc) Add better support for multiple story references via `StoryReferenceList`. 
-* (improvement) Improve exception message for better tracing/debugging when hydration of a `Story` fails due to invalid data.
+* (bc) Add better support for multiple story references via `StoryReferenceList`.
 
 
 1.5.0